requirements.txt
Pillow = ">=4.3.0"
streamlit = ">=1.20.0"